pyproject.toml
[project]
name = "financial-agent"
version = "0.1.0"
description = "LCL bank statement importer and AI-powered financial chat app"
authors = [
    {name = "Hassan Alhajj"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "openai>=1.0",
    "pandas>=2.0",
    "python-dotenv>=1.0",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
